Bugfix: fixed a issue where the category node for products was empty
Magento 2.2.5 introduced parallel indexers which caused the catalog-product
relationship to be saved to different table for each storeview. The table name
used in the query to retrieve those nodes had to be changed so it would pick the
appropiate table for the collection.

The table name is now generated by using the store id of the current collection.